STYLE : 아키텍쳐 구성
MVC 패턴으로 구성하고 싶었는데 뭔가 MVC 패턴 보다는 아키텍쳐 설계에 더
가까운거 같다...
